requirements.txt
torch
torchvision
pillow
fastapi
uvicorn
python-multipart
streamlit
requests
